// api/src/modules/business/constants/service-names.ts
export enum ServiceNames {
  AGENCY = 'BUSINESS/AGENCY_SERVICE',
  CANDIDATE = 'BUSINESS/CANDIDATE_SERVICE',
  DOER = 'BUSINESS/DOER_SERVICE',
  MISSION = 'BUSINESS/MISSION_SERVICE',
  QUESTION = 'BUSINESS/QUESTION_SERVICE',
  QUIZZ = 'BUSINESS/QUIZZ_SERVICE',
  QUIZZ_SHEET = 'BUSINESS/QUIZZ_SHEET_SERVICE',
  TIME_TABLE = 'BUSINESS/TIME_TABLE_SERVICE',
}