// api/src/modules/business/constants/use-case-names.ts
export enum UseCaseNames {
  CHECKOUT_FULFILL_UNLOCK_DOERS = 'BUSINESS/CHECKOUT_FULFILL_UNLOCK_DOERS_USE_CASE',
  CHECKOUT_SESSION_CREATE = 'BUSINESS/CHECKOUT_SESSION_CREATE_USE_CASE',
  CREATE_QUIZZ = 'BUSINESS/CREATE_QUIZZ_USE_CASE',
  DOER_VIDEO_CV_INIT_BATCH = 'BUSINESS/DOER_VIDEO_CV_INIT_BATCH_USE_CASE',
  MISSION_SET_FAVORITE_DOER = 'BUSINESS/MISSION_SET_FAVORITE_DOER_USE_CASE',
  MISSION_SET_SEEN_CANDIDATE = 'BUSINESS/MISSION_SET_SEEN_CANDIDATE_USE_CASE',
  MISSION_TIME_TABLE_UPDATE = 'BUSINESS/MISSION_TIME_TABLE_UPDATE_USE_CASE',
  MISSION_UPDATE = 'BUSINESS/MISSION_UPDATE_USE_CASE',
  MISSION_VALIDATE = 'BUSINESS/MISSION_VALIDATE_USE_CASE',
  QUIZZ_UPDATE = 'BUSINESS/QUIZZ_UPDATE_USE_CASE',
}